Refactor/Bugfix/Major: Ensure linearity of msgs
We were running into a bunch of issues before where we would get data about unknown agents.

The fundamental issue was that messages were not always arriving in order, and were being immediately executed.

With this new system, each message knows which message came before it.  Messages are not executed until the previous message has been received and executed.

This ensures that all updates play in a linear fashion.