Feature/Minor: Freeze static, config-like objects
This isn't to solve any problem in particular.  It's just that reasoning about mutable/immutable state is hard in a language like this.  `const` isn't really the same sort of guarantee in this language that it is in something like Haskell, or even Scala.  Mutability is pervasive and insidious (particularly, with regards to `Object`s), so having some sort of small guarantee that these objects won't be changed from beneath might feet, and, thus, can't be the source of state-management bugs, is slightly reassuring.