Optimization: Use faster heuristic for micro-burst branching
It takes an order of magnitude more time to serialize the message a second time than it does to stringify the JSON and see if it's small.  So let's just stringify.

This strategy *does* bear the caveat of skipping optimizations on larger messages, but I've only really seen that affect 'initial-model' messages.  And those were *already* like being hit with a sledgehammer, so... meh!