Optimization: Delay `BurstQueue.run` until right before expecting BQ msgs
This is pretty minor.  But it's kind of annoying that we were starting the burst queue up, only to then reject the client for username/password/capacity issues, leaving the BQ looping hopelessly in the background.

The thought process before was 'better safe (running the BQ early) than sorry (dropping messages)', but I think the client really does have time to do it this way.