Bugfix(?): Properly invert bitmasks
The uncertainty in my commit tag comes from the fact that... I'm not really clear on how this code was functioning at all before, because it sure wasn't doing what I thought it was!  That negation operator was actually making the number negative!  I just wanted it to flip every bit (as if it could actually know how many bits I wanted flipped in each case...).

Well... it wasn't doing that.  And, still, the code worked somehow.  So is this really fixing anything?  Unclear.