Bugfix: Refresh frame on joiner disconnect
This is being done instead of loading up a fake model.  The process of loading the fake model could take a long time to complete and put the next activity connected to into a weird, inconsistent state.  Specifically, this was manifesting as using the fake model's highlighting color when connecting to a new activity (on uncommon, random occasion).