Feature: Support Firefox beta release
This is a bit of a misnomer, I guess.  Firefox beta now supports *us*, given the resolution of Firefox bug 1247687.

However, we were incompatible with Firefox's implementation of WebRTC data channels, somewhat due to how Chromium is handling WebRTC issue 2276.  Essentially, Chromium defaults to sending data as `ArrayBuffer`s, even though the spec (and Firefox's implementation) say that the default is to send `Blob`s. (See the portion of the spec on DOM > DataChannel > binaryType)

`ArrayBuffer`s are great, but we need to explicitly opt into using them, in order for Firefox to do the right thing.