Revert "UX: Rebrand 'Authoring' as 'Author' in topbar"
This reverts commit bef8267b684aeb4b961dac671c0a2c19a3c6037b.

And it adds a couple of changes, beyond that, concerning this wording.

Reasoning: I liked "author", because it was imperative (like "host" and
"join") and shorter in length.  But it's actually a little confusing,
because you could mistake it for meaning that clicking the link would
bring you to a page for learning about the author of HNW.