Optimization: Make joiners get role data from web server
This optimization might not be worth the added complexity.  In fact, it is *probably* not worth it.  But it *might* be worth it, so I'm giving it a shot.  In particular, I'm trying here to alleviate performance woes when, say, a whole class of 100 people join a session all at once.

The idea is that the role data is statically known, at the time that the model is launched.  If we have a million clients connect to the session, we don't actually need to serialize that data a million times, nor have the host use their bandwidth on sending it.  This data makes up about 6-10% of an 'initial-model' message (the largest message type, by far), and my *hope* is that reducing the amount of CPU usage dedicated to these big messages will alleviate some of the out-of-order jitteriness of message output when there are a ton of clients.

These changes cause the role data to be uploaded to the web server when
the host launches the model.  Then, the joiner acquires the role data
over HTTP when they open communications with the host, and merge it into
the 'initial-model' message when that arrives.