Optimization: Queue up 'intial-model' messages
The 'initial-model' messages are, by far, the largest messages that we send.  When a host seems a huge influx of people connecting to an activity, it can be difficult to fetch and serialize all of those messages, while also updating all of the already-connected clients.  It leads to lagginess and failed attempts to connect.  So, instead, only allow one client to be connecting at a time.  The rest can sit in limbo.